+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Plugin 🕹️
 - Fix package_size not working for files
-- Add delete_unused_icons.py
+- Add delete_unused_icons
+- Fix error when Preferences 'theme' is set to 'auto'
+- Add system_theme, macOS
 
 ## [0.4.1] - 2024-12-15